Speculative caching of object properties
When StorageFactory is asked for properties of some object, we retrieve
them not only for the object in question, but for all its siblings as
well. Under assumption that it's likely those values will be requested
by the initiator shortly, the mass caching should improve performance
by minimizing access to Tracker or potential other slow storage that
might be added in the future.